Updating inline RAML examples
The README was out of date with the current version of the `raml-js-parser` hosted on `npmjs.org` (which matches the Github repo's version number, so I assume they're synced). The parser requires a a `'#%RAML 0.8'` string in the first line. It also threw the following once the version string was added:

```
Error parsing: while validating resources
property: 'name' is invalid in a resource
```

By changing the resource name to the explicit value (rather than the key-value pair that seems to be in the current docs), it seems to parse correctly as valid RAML.